+++ CuteDialog/src/main/java/io/github/cutelibs/cutedialog/CuteDialog.java
@@ -18,7 +18,18 @@ import androidx.core.content.ContextCompat;
 
 import io.github.cutelibs.cutedialog.databinding.CuteDialogMainLayoutBinding;
 
-public class CuteDialog {
+public class CuteDialog extends Dialog {
+
+    public static final int POSITION_CENTER = 1;
+    public static final int POSITION_TOP = 2;
+    public static final int POSITION_BOTTOM = 3;
+    public static final int STYLE_NORMAL = 1;
+    public static final int STYLE_BOLD = 2;
+    public static final int STYLE_ITALIC = 3;
+    public static final int STYLE_BOLD_ITALIC = 4;
+    public static final int HEADER_ICON = 1;
+    public static final int HEADER_IMAGE = 2;
+    public static final int HEADER_ANIMATION = 3;
 
     // Default Value
     private static final int WHOLE_BACKGROUND_COLOR_DEFAULT = Color.parseColor("#FFFFFF");
@@ -39,12 +50,513 @@ public class CuteDialog {
     private static final String NEGATIVE_BUTTON_TEXT_DEFAULT = "No";
     private static final String TITLE_TEXT_DEFAULT = "";
     private static final String DESC_TEXT_DEFAULT = "";
+    private final Context context;
+    CuteDialogMainLayoutBinding binding;
+    private int HEADER_CHOOSER = 1;
+
+    @Deprecated
+    public CuteDialog(Context context) {
+        super(context);
+        this.context = context;
+        binding = CuteDialogMainLayoutBinding.inflate(LayoutInflater.from(context));
+        requestWindowFeature(Window.FEATURE_NO_TITLE);
+        getWindow().setBackgroundDrawableResource(android.R.color.transparent);
+        getWindow().setLayout(FrameLayout.LayoutParams.MATCH_PARENT, FrameLayout.LayoutParams.WRAP_CONTENT);
+        setContentView(binding.getRoot());
+        init();
+
+    }
+
+    private void init() {
+        // whole card design
+        getWindow().setGravity(Gravity.CENTER);
+        setCancelable(true);
+        binding.wholeCard.setCardBackgroundColor(WHOLE_BACKGROUND_COLOR_DEFAULT);
+        binding.wholeCard.setRadius(WHOLE_CORNER_RADIUS_DEFAULT * 4);
+
+        // padding
+        binding.padding1.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+        binding.padding2.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 2));
+        binding.padding3.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+        binding.padding4.setLayoutParams(new LinearLayout.LayoutParams(WHOLE_PADDING_DEFAULT * 4, ViewGroup.LayoutParams.WRAP_CONTENT));
+        binding.padding5.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+        if (HEADER_CHOOSER == 1) {
+            binding.padding6.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+        } else {
+            binding.padding6.setVisibility(View.GONE);
+        }
+
+        // close icon
+        binding.closeIcon.setImageResource(R.drawable.cute_dialog_close_icon);
+        binding.closeIcon.setColorFilter(CLOSE_ICON_COLOR_DEFAULT);
+        binding.closeIcon.getLayoutParams().height = CLOSE_ICON_SIZE_DEFAULT * 2;
+        binding.closeIcon.getLayoutParams().width = CLOSE_ICON_SIZE_DEFAULT * 2;
+
+        // header visibility
+        binding.mainIcon.setVisibility(View.VISIBLE);
+        binding.mainImage.setVisibility(View.GONE);
+        binding.mainAnimation.setVisibility(View.GONE);
+
+        // title Text
+        binding.titleText.setText(TITLE_TEXT_DEFAULT);
+        binding.titleText.setTextSize(TypedValue.COMPLEX_UNIT_SP, TITLE_TEXT_SIZE_DEFAULT);
+        binding.titleText.setTextColor(TITLE_TEXT_COLOR_DEFAULT);
+        binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD);
+
+        // description Text
+        binding.descText.setText(DESC_TEXT_DEFAULT);
+        binding.descText.setTextSize(TypedValue.COMPLEX_UNIT_SP, DESC_TEXT_SIZE_DEFAULT);
+        binding.descText.setTextColor(DESC_TEXT_COLOR_DEFAULT);
+        binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+
+        // positive button style
+        binding.positiveButton.setRadius(BUTTON_CORNER_RADIUS_DEFAULT * 4);
+        binding.positiveButton.setCardBackgroundColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+        binding.positiveButton.setStrokeColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+        binding.positiveButton.setStrokeWidth(1);
+
+        // negative button style
+        binding.negativeButton.setRadius(BUTTON_CORNER_RADIUS_DEFAULT * 4);
+        binding.negativeButton.setCardBackgroundColor(WHOLE_BACKGROUND_COLOR_DEFAULT);
+        binding.negativeButton.setStrokeColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+        binding.negativeButton.setStrokeWidth(1);
+
+        // button text
+        binding.positiveText.setTextSize(TypedValue.COMPLEX_UNIT_SP, BUTTON_TEXT_SIZE_DEFAULT);
+        binding.negativeText.setTextSize(TypedValue.COMPLEX_UNIT_SP, BUTTON_TEXT_SIZE_DEFAULT);
+        binding.positiveText.setText(POSITIVE_BUTTON_TEXT_DEFAULT);
+        binding.positiveText.setTextColor(POSITIVE_BUTTON_TEXT_COLOR_DEFAULT);
+        binding.negativeText.setText(NEGATIVE_BUTTON_TEXT_DEFAULT);
+        binding.negativeText.setTextColor(NEGATIVE_BUTTON_TEXT_COLOR_DEFAULT);
+
+        // button click
+        binding.positiveButton.setOnClickListener(v -> dismiss());
+        binding.negativeButton.setOnClickListener(v -> dismiss());
+        binding.closeIcon.setOnClickListener(v -> dismiss());
+
+    }
+
+    @Deprecated
+    public CuteDialog setDialogStyle(int bgColor, int cornerRadius, int dialogPosition, int padding) {
+
+        // set dialog position
+        if (dialogPosition == 1) {
+            getWindow().setGravity(Gravity.CENTER);
+        } else if (dialogPosition == 2) {
+            getWindow().setGravity(Gravity.TOP);
+        } else if (dialogPosition == 3) {
+            getWindow().setGravity(Gravity.BOTTOM);
+        } else {
+            getWindow().setGravity(Gravity.CENTER);
+        }
+
+        // wholeCard background
+        if (bgColor != 0) {
+            try {
+                binding.wholeCard.setCardBackgroundColor(ContextCompat.getColor(context, bgColor));
+            } catch (Resources.NotFoundException e) {
+                binding.wholeCard.setCardBackgroundColor(bgColor);
+            }
+        }
+
+        // wholeCard corner radius
+        if (cornerRadius != 0) {
+            binding.wholeCard.setRadius(cornerRadius * 4);
+        }
+
+        // wholeCard padding
+        if (padding != 0) {
+            binding.padding1.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+            binding.padding2.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 2));
+            binding.padding3.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+            binding.padding4.setLayoutParams(new LinearLayout.LayoutParams(padding * 4, ViewGroup.LayoutParams.WRAP_CONTENT));
+            binding.padding5.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+            if (HEADER_CHOOSER == 1) {
+                binding.padding6.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+            } else {
+                binding.padding6.setVisibility(View.GONE);
+            }
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog isCancelable(boolean cancelable) {
+        setCancelable(cancelable);
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setCloseIconStyle(int iconResID, int sizeInDP, int colorOfIcon) {
+
+        // closeIcon
+        if (iconResID != 0) {
+            try {
+                binding.closeIcon.setImageResource(iconResID);
+            } catch (Resources.NotFoundException e) {
+                binding.closeIcon.setImageResource(R.drawable.cute_dialog_close_icon);
+            }
+        }
+
+        // closeIcon color
+        if (colorOfIcon != 0) {
+            try {
+                binding.closeIcon.setColorFilter(ContextCompat.getColor(context, colorOfIcon));
+            } catch (Resources.NotFoundException e) {
+                binding.closeIcon.setColorFilter(colorOfIcon);
+            }
+        }
+
+        // closeIcon size
+        if (sizeInDP != 0) {
+            binding.closeIcon.getLayoutParams().height = sizeInDP * 2;
+            binding.closeIcon.getLayoutParams().width = sizeInDP * 2;
+            binding.closeIcon.requestLayout();
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setHeader(int chooser) {
+        HEADER_CHOOSER = chooser;
+
+        if (chooser == 2) {
+            binding.mainIcon.setVisibility(View.GONE);
+            binding.mainImage.setVisibility(View.VISIBLE);
+            binding.mainAnimation.setVisibility(View.GONE);
+            binding.padding6.setVisibility(View.GONE);
+        } else if (chooser == 3) {
+            binding.mainIcon.setVisibility(View.GONE);
+            binding.mainImage.setVisibility(View.GONE);
+            binding.mainAnimation.setVisibility(View.VISIBLE);
+            binding.padding6.setVisibility(View.GONE);
+        } else {
+            binding.mainIcon.setVisibility(View.VISIBLE);
+            binding.mainImage.setVisibility(View.GONE);
+            binding.mainAnimation.setVisibility(View.GONE);
+            binding.padding6.setVisibility(View.VISIBLE);
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setHeaderIcon(int icon) {
+        try {
+            binding.mainIcon.setImageResource(icon);
+        } catch (Resources.NotFoundException e) {
+            e.printStackTrace();
+        }
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setHeaderImage(int image) {
+        try {
+            binding.mainImage.setImageResource(image);
+        } catch (Resources.NotFoundException e) {
+            e.printStackTrace();
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setHeaderAnimation(int animation) {
+        try {
+            binding.mainAnimation.setAnimation(animation);
+        } catch (Resources.NotFoundException e) {
+            e.printStackTrace();
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setTitle(String titleText, int textSizeInSP, int textColor, int textStyle) {
+        binding.titleText.setText(titleText);
+
+        if (textSizeInSP != 0) {
+            binding.titleText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+        }
+
+        // title Text color
+        if (textColor != 0) {
+            try {
+                binding.titleText.setTextColor(ContextCompat.getColor(context, textColor));
+            } catch (Resources.NotFoundException e) {
+                binding.titleText.setTextColor(textColor);
+            }
+        }
+
+        // title text style
+        if (textStyle == 1) {
+            binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.NORMAL);
+        } else if (textStyle == 2) {
+            binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD);
+        } else if (textStyle == 3) {
+            binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.ITALIC);
+        } else if (textStyle == 4) {
+            binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD_ITALIC);
+        } else {
+            binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.NORMAL);
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setDesc(String descText, int textSizeInSP, int textColor, int textStyle) {
+
+        binding.descText.setText(descText);
+
+        if (textSizeInSP != 0) {
+            binding.descText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+        }
+
+        // desc Text color
+        if (textColor != 0) {
+            try {
+                binding.descText.setTextColor(ContextCompat.getColor(context, textColor));
+            } catch (Resources.NotFoundException e) {
+                binding.descText.setTextColor(textColor);
+            }
+        }
+
+        // desc text style
+        if (textStyle == 1) {
+            binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+        } else if (textStyle == 2) {
+            binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.BOLD);
+        } else if (textStyle == 3) {
+            binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.ITALIC);
+        } else if (textStyle == 4) {
+            binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.BOLD_ITALIC);
+        } else {
+            binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setPositiveButtonStyle(int radiusOfButton, int bgColorOfButton, int borderColor, int borderWidth, int textSizeInSP) {
+
+        if (radiusOfButton != 0) {
+            binding.positiveButton.setRadius(radiusOfButton * 4);
+        }
+
+        if (bgColorOfButton != 0) {
+            try {
+                binding.positiveButton.setCardBackgroundColor(ContextCompat.getColor(context, bgColorOfButton));
+            } catch (Resources.NotFoundException e) {
+                binding.positiveButton.setCardBackgroundColor(bgColorOfButton);
+            }
+        }
+
+        if (borderColor != 0) {
+            try {
+                binding.positiveButton.setStrokeColor(ContextCompat.getColor(context, borderColor));
+            } catch (Resources.NotFoundException e) {
+                binding.positiveButton.setStrokeColor(borderColor);
+            }
+        } else {
+            if (bgColorOfButton != 0) {
+                try {
+                    binding.positiveButton.setStrokeColor(ContextCompat.getColor(context, bgColorOfButton));
+                } catch (Resources.NotFoundException e) {
+                    binding.positiveButton.setStrokeColor(bgColorOfButton);
+                }
+            }
+        }
+
+        if (borderWidth != 0) {
+            binding.positiveButton.setStrokeWidth(borderWidth);
+        }
+
+        if (textSizeInSP != 0) {
+            binding.positiveText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setNegativeButtonStyle(int radiusOfButton, int bgColorOfButton, int borderColor, int borderWidth, int textSizeInSP) {
+
+        if (radiusOfButton != 0) {
+            binding.negativeButton.setRadius(radiusOfButton * 4);
+
+        }
+
+        if (bgColorOfButton != 0) {
+            try {
+                binding.negativeButton.setCardBackgroundColor(ContextCompat.getColor(context, bgColorOfButton));
+            } catch (Resources.NotFoundException e) {
+                binding.negativeButton.setCardBackgroundColor(bgColorOfButton);
+            }
+        }
+
+        if (borderColor != 0) {
+            try {
+                binding.negativeButton.setStrokeColor(ContextCompat.getColor(context, borderColor));
+            } catch (Resources.NotFoundException e) {
+                binding.negativeButton.setStrokeColor(borderColor);
+            }
+        } else {
+            if (bgColorOfButton != 0) {
+                try {
+                    binding.negativeButton.setStrokeColor(ContextCompat.getColor(context, bgColorOfButton));
+                } catch (Resources.NotFoundException e) {
+                    binding.negativeButton.setStrokeColor(bgColorOfButton);
+                }
+            }
+        }
+
+        if (borderWidth != 0) {
+            binding.negativeButton.setStrokeWidth(borderWidth);
+        }
+
+
+        if (textSizeInSP != 0) {
+            binding.negativeText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+        }
+
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setPositiveButtonText(String positiveText, int textColor, int textStyle) {
+        binding.positiveText.setText(positiveText);
+        if (textColor != 0) {
+            try {
+                binding.positiveText.setTextColor(ContextCompat.getColor(context, textColor));
+            } catch (Resources.NotFoundException e) {
+                binding.positiveText.setTextColor(textColor);
+            }
+        }
+
+        // positive text style
+        if (textStyle == 1) {
+            binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.NORMAL);
+        } else if (textStyle == 2) {
+            binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.BOLD);
+        } else if (textStyle == 3) {
+            binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.ITALIC);
+        } else if (textStyle == 4) {
+            binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.BOLD_ITALIC);
+        } else {
+            binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.NORMAL);
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setNegativeButtonText(String negativeText, int textColor, int textStyle) {
+
+        binding.negativeText.setText(negativeText);
+        if (textColor != 0) {
+            try {
+                binding.negativeText.setTextColor(ContextCompat.getColor(context, textColor));
+            } catch (Resources.NotFoundException e) {
+                binding.negativeText.setTextColor(textColor);
+            }
+        }
+
+        // negative text style
+        if (textStyle == 1) {
+            binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.NORMAL);
+        } else if (textStyle == 2) {
+            binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.BOLD);
+        } else if (textStyle == 3) {
+            binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.ITALIC);
+        } else if (textStyle == 4) {
+            binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.BOLD_ITALIC);
+        } else {
+            binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.NORMAL);
+        }
+
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setVisibilityOptions(boolean hideCloseIcon, boolean hideTitle, boolean hideDesc, boolean hidePositiveButton, boolean hideNegativeButton) {
+
+        if (hideCloseIcon) {
+            binding.closeIcon.setVisibility(View.GONE);
+        } else {
+            binding.closeIcon.setVisibility(View.VISIBLE);
+        }
+
+        if (hideTitle) {
+            binding.titleText.setVisibility(View.GONE);
+        } else {
+            binding.titleText.setVisibility(View.VISIBLE);
+        }
+
+        if (hideDesc) {
+            binding.descText.setVisibility(View.GONE);
+        } else {
+            binding.descText.setVisibility(View.VISIBLE);
+        }
+
+        if (hidePositiveButton) {
+            binding.positiveButton.setVisibility(View.GONE);
+        } else {
+            binding.positiveButton.setVisibility(View.VISIBLE);
+        }
+
+        if (hideNegativeButton) {
+            binding.negativeButton.setVisibility(View.GONE);
+        } else {
+            binding.negativeButton.setVisibility(View.VISIBLE);
+        }
+
+        if (hideNegativeButton || hidePositiveButton) {
+            binding.padding4.setVisibility(View.GONE);
+        } else {
+            binding.padding4.setVisibility(View.VISIBLE);
+        }
+
+        return this;
+
+    }
+
+    @Deprecated
+    public CuteDialog setPositiveButtonListener(View.OnClickListener listener) {
+        binding.positiveButton.setOnClickListener(v -> {
+            listener.onClick(v);
+            dismiss();
+        });
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setNegativeButtonListener(View.OnClickListener listener) {
+        binding.negativeButton.setOnClickListener(v -> {
+            listener.onClick(v);
+            dismiss();
+        });
+        return this;
+    }
+
+    @Deprecated
+    public CuteDialog setCloseListener(View.OnClickListener listener) {
+        binding.closeIcon.setOnClickListener(v -> {
+            listener.onClick(v);
+            dismiss();
+        });
+        return this;
+    }
 
 
     public static class withIcon extends Dialog {
 
         Context context;
-
         CuteDialogMainLayoutBinding binding;
 
         public withIcon(Context context) {
@@ -125,8 +637,49 @@ public class CuteDialog {
 
         }
 
+        public withIcon setDialogBackgroundColor(int bgColor) {
+            if (bgColor != 0) {
+                try {
+                    binding.wholeCard.setCardBackgroundColor(ContextCompat.getColor(context, bgColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.wholeCard.setCardBackgroundColor(bgColor);
+                }
+            }
+            return this;
+        }
 
-        // set Icon
+        public withIcon setDialogRadius(int cornerRadius) {
+            if (cornerRadius != 0) {
+                binding.wholeCard.setRadius(cornerRadius * 4);
+            }
+            return this;
+        }
+
+        public withIcon setDialogPosition(int dialogPosition) {
+            if (dialogPosition == 1) {
+                getWindow().setGravity(Gravity.CENTER);
+            } else if (dialogPosition == 2) {
+                getWindow().setGravity(Gravity.TOP);
+            } else if (dialogPosition == 3) {
+                getWindow().setGravity(Gravity.BOTTOM);
+            } else {
+                getWindow().setGravity(Gravity.CENTER);
+            }
+            return this;
+        }
+
+        public withIcon setPadding(int padding) {
+            if (padding != 0) {
+                binding.padding1.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding2.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 2));
+                binding.padding3.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding4.setLayoutParams(new LinearLayout.LayoutParams(padding * 4, ViewGroup.LayoutParams.WRAP_CONTENT));
+                binding.padding5.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding6.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+            }
+            return this;
+        }
+
         public withIcon setIcon(int icon) {
             try {
                 binding.mainIcon.setImageResource(icon);
@@ -136,15 +689,11 @@ public class CuteDialog {
             return this;
         }
 
-
-        // set Cancelable
         public withIcon isCancelable(boolean cancelable) {
             setCancelable(cancelable);
             return this;
         }
 
-
-        // Set Title Text and Style
         public withIcon setTitle(String string) {
             binding.titleText.setText(string);
             return this;
@@ -183,8 +732,6 @@ public class CuteDialog {
             return this;
         }
 
-
-        // Set Description Text and Style
         public withIcon setDescription(String descText) {
             binding.descText.setText(descText);
             return this;
@@ -224,8 +771,6 @@ public class CuteDialog {
             return this;
         }
 
-
-        // Set Positive Text and Style
         public withIcon setPositiveButtonText(String string, View.OnClickListener listener) {
             binding.positiveText.setText(string);
             binding.positiveButton.setOnClickListener(v -> {
@@ -269,8 +814,6 @@ public class CuteDialog {
             return this;
         }
 
-
-        // Set Positive Button Style
         public withIcon setPositiveButtonColor(int bgColorOfButton) {
             if (bgColorOfButton != 0) {
                 try {
@@ -309,8 +852,6 @@ public class CuteDialog {
             return this;
         }
 
-
-        // Set Negative Text and Style
         public withIcon setNegativeButtonText(String string, View.OnClickListener listener) {
             binding.negativeText.setText(string);
             binding.negativeButton.setOnClickListener(v -> {
@@ -353,8 +894,6 @@ public class CuteDialog {
             return this;
         }
 
-
-        // Set Negative Button Style
         public withIcon setNegativeButtonColor(int bgColorOfButton) {
             if (bgColorOfButton != 0) {
                 try {
@@ -394,8 +933,6 @@ public class CuteDialog {
             return this;
         }
 
-
-        // hide options
         public withIcon hideCloseIcon(boolean bool) {
             if (bool) {
                 binding.closeIcon.setVisibility(View.GONE);
@@ -448,8 +985,6 @@ public class CuteDialog {
 
         }
 
-
-        // close icon
         public withIcon setCloseIconListener(View.OnClickListener listener) {
             binding.closeIcon.setOnClickListener(v -> {
                 listener.onClick(v);
@@ -479,6 +1014,975 @@ public class CuteDialog {
         }
 
         public withIcon setCloseIconColor(int colorOfIcon) {
+            if (colorOfIcon != 0) {
+                try {
+                    binding.closeIcon.setColorFilter(ContextCompat.getColor(context, colorOfIcon));
+                } catch (Resources.NotFoundException e) {
+                    binding.closeIcon.setColorFilter(colorOfIcon);
+                }
+            }
+            return this;
+        }
+
+    }
+
+    public static class withImage extends Dialog {
+
+        Context context;
+
+        CuteDialogMainLayoutBinding binding;
+
+        public withImage(Context context) {
+            super(context);
+            this.context = context;
+            binding = CuteDialogMainLayoutBinding.inflate(LayoutInflater.from(context));
+            requestWindowFeature(Window.FEATURE_NO_TITLE);
+            getWindow().setBackgroundDrawableResource(android.R.color.transparent);
+            getWindow().setLayout(FrameLayout.LayoutParams.MATCH_PARENT, FrameLayout.LayoutParams.WRAP_CONTENT);
+            setContentView(binding.getRoot());
+            inits();
+
+        }
+
+        private void inits() {
+            // whole card design
+            getWindow().setGravity(Gravity.CENTER);
+            setCancelable(true);
+            binding.wholeCard.setCardBackgroundColor(WHOLE_BACKGROUND_COLOR_DEFAULT);
+            binding.wholeCard.setRadius(WHOLE_CORNER_RADIUS_DEFAULT * 4);
+
+            // padding
+            binding.padding1.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+            binding.padding2.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 2));
+            binding.padding3.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+            binding.padding4.setLayoutParams(new LinearLayout.LayoutParams(WHOLE_PADDING_DEFAULT * 4, ViewGroup.LayoutParams.WRAP_CONTENT));
+            binding.padding5.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+            binding.padding6.setVisibility(View.GONE);
+
+            // close icon
+            binding.closeIcon.setImageResource(R.drawable.cute_dialog_close_icon);
+            binding.closeIcon.setColorFilter(CLOSE_ICON_COLOR_DEFAULT);
+            binding.closeIcon.getLayoutParams().height = CLOSE_ICON_SIZE_DEFAULT * 2;
+            binding.closeIcon.getLayoutParams().width = CLOSE_ICON_SIZE_DEFAULT * 2;
+
+            // header visibility
+            binding.mainIcon.setVisibility(View.GONE);
+            binding.mainImage.setVisibility(View.VISIBLE);
+            binding.mainAnimation.setVisibility(View.GONE);
+
+            // title Text
+            binding.titleText.setText(TITLE_TEXT_DEFAULT);
+            binding.titleText.setTextSize(TypedValue.COMPLEX_UNIT_SP, TITLE_TEXT_SIZE_DEFAULT);
+            binding.titleText.setTextColor(TITLE_TEXT_COLOR_DEFAULT);
+            binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD);
+
+            // description Text
+            binding.descText.setText(DESC_TEXT_DEFAULT);
+            binding.descText.setTextSize(TypedValue.COMPLEX_UNIT_SP, DESC_TEXT_SIZE_DEFAULT);
+            binding.descText.setTextColor(DESC_TEXT_COLOR_DEFAULT);
+            binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+
+            // positive button style
+            binding.positiveButton.setRadius(BUTTON_CORNER_RADIUS_DEFAULT * 4);
+            binding.positiveButton.setCardBackgroundColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+            binding.positiveButton.setStrokeColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+            binding.positiveButton.setStrokeWidth(1);
+
+            // negative button style
+            binding.negativeButton.setRadius(BUTTON_CORNER_RADIUS_DEFAULT * 4);
+            binding.negativeButton.setCardBackgroundColor(WHOLE_BACKGROUND_COLOR_DEFAULT);
+            binding.negativeButton.setStrokeColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+            binding.negativeButton.setStrokeWidth(1);
+
+            // button text
+            binding.positiveText.setTextSize(TypedValue.COMPLEX_UNIT_SP, BUTTON_TEXT_SIZE_DEFAULT);
+            binding.negativeText.setTextSize(TypedValue.COMPLEX_UNIT_SP, BUTTON_TEXT_SIZE_DEFAULT);
+            binding.positiveText.setText(POSITIVE_BUTTON_TEXT_DEFAULT);
+            binding.positiveText.setTextColor(POSITIVE_BUTTON_TEXT_COLOR_DEFAULT);
+            binding.negativeText.setText(NEGATIVE_BUTTON_TEXT_DEFAULT);
+            binding.negativeText.setTextColor(NEGATIVE_BUTTON_TEXT_COLOR_DEFAULT);
+
+            // button click
+            binding.positiveButton.setOnClickListener(v -> dismiss());
+            binding.negativeButton.setOnClickListener(v -> dismiss());
+            binding.closeIcon.setOnClickListener(v -> dismiss());
+
+        }
+
+        public withImage setDialogBackgroundColor(int bgColor) {
+            if (bgColor != 0) {
+                try {
+                    binding.wholeCard.setCardBackgroundColor(ContextCompat.getColor(context, bgColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.wholeCard.setCardBackgroundColor(bgColor);
+                }
+            }
+            return this;
+        }
+
+        public withImage setDialogRadius(int cornerRadius) {
+            if (cornerRadius != 0) {
+                binding.wholeCard.setRadius(cornerRadius * 4);
+            }
+            return this;
+        }
+
+        public withImage setDialogPosition(int dialogPosition) {
+            if (dialogPosition == 1) {
+                getWindow().setGravity(Gravity.CENTER);
+            } else if (dialogPosition == 2) {
+                getWindow().setGravity(Gravity.TOP);
+            } else if (dialogPosition == 3) {
+                getWindow().setGravity(Gravity.BOTTOM);
+            } else {
+                getWindow().setGravity(Gravity.CENTER);
+            }
+            return this;
+        }
+
+        public withImage setPadding(int padding) {
+            if (padding != 0) {
+                binding.padding1.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding2.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 2));
+                binding.padding3.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding4.setLayoutParams(new LinearLayout.LayoutParams(padding * 4, ViewGroup.LayoutParams.WRAP_CONTENT));
+                binding.padding5.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding6.setVisibility(View.GONE);
+            }
+            return this;
+        }
+
+        public withImage setHeaderImage(int image) {
+            try {
+                binding.mainImage.setImageResource(image);
+            } catch (Resources.NotFoundException e) {
+                e.printStackTrace();
+            }
+
+            return this;
+        }
+
+        public withImage isCancelable(boolean cancelable) {
+            setCancelable(cancelable);
+            return this;
+        }
+
+        public withImage setTitle(String string) {
+            binding.titleText.setText(string);
+            return this;
+        }
+
+        public withImage setTitleTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.titleText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+        public withImage setTitleTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.titleText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.titleText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withImage setTitleTextStyle(int textStyle) {
+            if (textStyle == 1) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.NORMAL);
+            }
+            return this;
+        }
+
+        public withImage setDescription(String descText) {
+            binding.descText.setText(descText);
+            return this;
+        }
+
+        public withImage setDescriptionTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.descText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+        public withImage setDescriptionTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.descText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.descText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withImage setDescriptionTextStyle(int textStyle) {
+            // desc text style
+            if (textStyle == 1) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+            }
+            return this;
+        }
+
+        public withImage setPositiveButtonText(String string, View.OnClickListener listener) {
+            binding.positiveText.setText(string);
+            binding.positiveButton.setOnClickListener(v -> {
+                listener.onClick(v);
+                dismiss();
+            });
+            return this;
+        }
+
+        public withImage setPositiveButtonTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.positiveText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.positiveText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withImage setPositiveButtonTextStyle(int textStyle) {
+            if (textStyle == 1) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.NORMAL);
+            }
+
+            return this;
+        }
+
+        public withImage setPositiveButtonTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.positiveText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+        public withImage setPositiveButtonColor(int bgColorOfButton) {
+            if (bgColorOfButton != 0) {
+                try {
+                    binding.positiveButton.setCardBackgroundColor(ContextCompat.getColor(context, bgColorOfButton));
+                } catch (Resources.NotFoundException e) {
+                    binding.positiveButton.setCardBackgroundColor(bgColorOfButton);
+                }
+            }
+            return this;
+        }
+
+        public withImage setPositiveButtonRadius(int radiusOfButton) {
+            if (radiusOfButton != 0) {
+                binding.positiveButton.setRadius(radiusOfButton * 4);
+            }
+            return this;
+        }
+
+        public withImage setPositiveButtonBorderColor(int borderColor) {
+            if (borderColor != 0) {
+                try {
+                    binding.positiveButton.setStrokeColor(ContextCompat.getColor(context, borderColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.positiveButton.setStrokeColor(borderColor);
+                }
+            } else {
+                binding.positiveButton.setStrokeColor(Color.TRANSPARENT);
+            }
+            return this;
+        }
+
+        public withImage setPositiveButtonBorderWidth(int borderWidth) {
+            if (borderWidth != 0) {
+                binding.positiveButton.setStrokeWidth(borderWidth);
+            }
+            return this;
+        }
+
+        public withImage setNegativeButtonText(String string, View.OnClickListener listener) {
+            binding.negativeText.setText(string);
+            binding.negativeButton.setOnClickListener(v -> {
+                listener.onClick(v);
+                dismiss();
+            });
+            return this;
+        }
+
+        public withImage setNegativeButtonTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.negativeText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.negativeText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withImage setNegativeButtonTextStyle(int textStyle) {
+            if (textStyle == 1) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.NORMAL);
+            }
+            return this;
+        }
+
+        public withImage setNegativeButtonTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.negativeText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+        public withImage setNegativeButtonColor(int bgColorOfButton) {
+            if (bgColorOfButton != 0) {
+                try {
+                    binding.negativeButton.setCardBackgroundColor(ContextCompat.getColor(context, bgColorOfButton));
+                } catch (Resources.NotFoundException e) {
+                    binding.negativeButton.setCardBackgroundColor(bgColorOfButton);
+                }
+            }
+            return this;
+        }
+
+        public withImage setNegativeButtonRadius(int radiusOfButton) {
+            if (radiusOfButton != 0) {
+                binding.negativeButton.setRadius(radiusOfButton * 4);
+
+            }
+            return this;
+        }
+
+        public withImage setNegativeButtonBorderColor(int borderColor) {
+            if (borderColor != 0) {
+                try {
+                    binding.negativeButton.setStrokeColor(ContextCompat.getColor(context, borderColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.negativeButton.setStrokeColor(borderColor);
+                }
+            } else {
+                binding.negativeButton.setStrokeColor(Color.TRANSPARENT);
+            }
+            return this;
+        }
+
+        public withImage setNegativeButtonBorderWidth(int borderWidth) {
+            if (borderWidth != 0) {
+                binding.negativeButton.setStrokeWidth(borderWidth);
+            }
+            return this;
+        }
+
+        public withImage hideCloseIcon(boolean bool) {
+            if (bool) {
+                binding.closeIcon.setVisibility(View.GONE);
+            } else {
+                binding.closeIcon.setVisibility(View.VISIBLE);
+            }
+            return this;
+
+        }
+
+        public withImage hideTitle(boolean bool) {
+            if (bool) {
+                binding.titleText.setVisibility(View.GONE);
+            } else {
+                binding.titleText.setVisibility(View.VISIBLE);
+            }
+            return this;
+        }
+
+        public withImage hideDescription(boolean bool) {
+            if (bool) {
+                binding.descText.setVisibility(View.GONE);
+            } else {
+                binding.descText.setVisibility(View.VISIBLE);
+            }
+            return this;
+        }
+
+        public withImage hidePositiveButton(boolean bool) {
+            if (bool) {
+                binding.positiveButton.setVisibility(View.GONE);
+                binding.padding4.setVisibility(View.GONE);
+            } else {
+                binding.positiveButton.setVisibility(View.VISIBLE);
+            }
+            return this;
+        }
+
+        public withImage hideNegativeButton(boolean bool) {
+
+            if (bool) {
+                binding.negativeButton.setVisibility(View.GONE);
+                binding.padding4.setVisibility(View.GONE);
+            } else {
+                binding.negativeButton.setVisibility(View.VISIBLE);
+            }
+
+
+            return this;
+
+        }
+
+        public withImage setCloseIconListener(View.OnClickListener listener) {
+            binding.closeIcon.setOnClickListener(v -> {
+                listener.onClick(v);
+                dismiss();
+            });
+            return this;
+        }
+
+        public withImage setCloseIcon(int iconResID) {
+            if (iconResID != 0) {
+                try {
+                    binding.closeIcon.setImageResource(iconResID);
+                } catch (Resources.NotFoundException e) {
+                    binding.closeIcon.setImageResource(R.drawable.cute_dialog_close_icon);
+                }
+            }
+            return this;
+        }
+
+        public withImage setCloseIconSize(int sizeInDP) {
+            if (sizeInDP != 0) {
+                binding.closeIcon.getLayoutParams().height = sizeInDP * 2;
+                binding.closeIcon.getLayoutParams().width = sizeInDP * 2;
+                binding.closeIcon.requestLayout();
+            }
+            return this;
+        }
+
+        public withImage setCloseIconColor(int colorOfIcon) {
+            if (colorOfIcon != 0) {
+                try {
+                    binding.closeIcon.setColorFilter(ContextCompat.getColor(context, colorOfIcon));
+                } catch (Resources.NotFoundException e) {
+                    binding.closeIcon.setColorFilter(colorOfIcon);
+                }
+            }
+            return this;
+        }
+
+
+    }
+
+    public static class withAnim extends Dialog {
+
+        Context context;
+
+        CuteDialogMainLayoutBinding binding;
+
+        public withAnim(Context context) {
+            super(context);
+            this.context = context;
+            binding = CuteDialogMainLayoutBinding.inflate(LayoutInflater.from(context));
+            requestWindowFeature(Window.FEATURE_NO_TITLE);
+            getWindow().setBackgroundDrawableResource(android.R.color.transparent);
+            getWindow().setLayout(FrameLayout.LayoutParams.MATCH_PARENT, FrameLayout.LayoutParams.WRAP_CONTENT);
+            setContentView(binding.getRoot());
+            inits();
+
+        }
+
+        private void inits() {
+            // whole card design
+            getWindow().setGravity(Gravity.CENTER);
+            setCancelable(true);
+            binding.wholeCard.setCardBackgroundColor(WHOLE_BACKGROUND_COLOR_DEFAULT);
+            binding.wholeCard.setRadius(WHOLE_CORNER_RADIUS_DEFAULT * 4);
+
+            // padding
+            binding.padding1.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+            binding.padding2.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 2));
+            binding.padding3.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+            binding.padding4.setLayoutParams(new LinearLayout.LayoutParams(WHOLE_PADDING_DEFAULT * 4, ViewGroup.LayoutParams.WRAP_CONTENT));
+            binding.padding5.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, WHOLE_PADDING_DEFAULT * 4));
+            binding.padding6.setVisibility(View.GONE);
+
+            // close icon
+            binding.closeIcon.setImageResource(R.drawable.cute_dialog_close_icon);
+            binding.closeIcon.setColorFilter(CLOSE_ICON_COLOR_DEFAULT);
+            binding.closeIcon.getLayoutParams().height = CLOSE_ICON_SIZE_DEFAULT * 2;
+            binding.closeIcon.getLayoutParams().width = CLOSE_ICON_SIZE_DEFAULT * 2;
+
+            // header visibility
+            binding.mainIcon.setVisibility(View.GONE);
+            binding.mainImage.setVisibility(View.GONE);
+            binding.mainAnimation.setVisibility(View.VISIBLE);
+
+            // title Text
+            binding.titleText.setText(TITLE_TEXT_DEFAULT);
+            binding.titleText.setTextSize(TypedValue.COMPLEX_UNIT_SP, TITLE_TEXT_SIZE_DEFAULT);
+            binding.titleText.setTextColor(TITLE_TEXT_COLOR_DEFAULT);
+            binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD);
+
+            // description Text
+            binding.descText.setText(DESC_TEXT_DEFAULT);
+            binding.descText.setTextSize(TypedValue.COMPLEX_UNIT_SP, DESC_TEXT_SIZE_DEFAULT);
+            binding.descText.setTextColor(DESC_TEXT_COLOR_DEFAULT);
+            binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+
+            // positive button style
+            binding.positiveButton.setRadius(BUTTON_CORNER_RADIUS_DEFAULT * 4);
+            binding.positiveButton.setCardBackgroundColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+            binding.positiveButton.setStrokeColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+            binding.positiveButton.setStrokeWidth(1);
+
+            // negative button style
+            binding.negativeButton.setRadius(BUTTON_CORNER_RADIUS_DEFAULT * 4);
+            binding.negativeButton.setCardBackgroundColor(WHOLE_BACKGROUND_COLOR_DEFAULT);
+            binding.negativeButton.setStrokeColor(BUTTON_BACKGROUND_COLOR_DEFAULT);
+            binding.negativeButton.setStrokeWidth(1);
+
+            // button text
+            binding.positiveText.setTextSize(TypedValue.COMPLEX_UNIT_SP, BUTTON_TEXT_SIZE_DEFAULT);
+            binding.negativeText.setTextSize(TypedValue.COMPLEX_UNIT_SP, BUTTON_TEXT_SIZE_DEFAULT);
+            binding.positiveText.setText(POSITIVE_BUTTON_TEXT_DEFAULT);
+            binding.positiveText.setTextColor(POSITIVE_BUTTON_TEXT_COLOR_DEFAULT);
+            binding.negativeText.setText(NEGATIVE_BUTTON_TEXT_DEFAULT);
+            binding.negativeText.setTextColor(NEGATIVE_BUTTON_TEXT_COLOR_DEFAULT);
+
+            // button click
+            binding.positiveButton.setOnClickListener(v -> dismiss());
+            binding.negativeButton.setOnClickListener(v -> dismiss());
+            binding.closeIcon.setOnClickListener(v -> dismiss());
+
+        }
+
+
+        // set Dialog Style
+        public withAnim setDialogBackgroundColor(int bgColor) {
+            if (bgColor != 0) {
+                try {
+                    binding.wholeCard.setCardBackgroundColor(ContextCompat.getColor(context, bgColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.wholeCard.setCardBackgroundColor(bgColor);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setDialogRadius(int cornerRadius) {
+            if (cornerRadius != 0) {
+                binding.wholeCard.setRadius(cornerRadius * 4);
+            }
+            return this;
+        }
+
+        public withAnim setDialogPosition(int dialogPosition) {
+            if (dialogPosition == 1) {
+                getWindow().setGravity(Gravity.CENTER);
+            } else if (dialogPosition == 2) {
+                getWindow().setGravity(Gravity.TOP);
+            } else if (dialogPosition == 3) {
+                getWindow().setGravity(Gravity.BOTTOM);
+            } else {
+                getWindow().setGravity(Gravity.CENTER);
+            }
+            return this;
+        }
+
+        public withAnim setPadding(int padding) {
+            if (padding != 0) {
+                binding.padding1.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding2.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 2));
+                binding.padding3.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding4.setLayoutParams(new LinearLayout.LayoutParams(padding * 4, ViewGroup.LayoutParams.WRAP_CONTENT));
+                binding.padding5.setLayoutParams(new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, padding * 4));
+                binding.padding6.setVisibility(View.GONE);
+            }
+            return this;
+        }
+
+
+        // set Animation
+        public withAnim setAnimation(int animation) {
+            try {
+                binding.mainAnimation.setAnimation(animation);
+            } catch (Resources.NotFoundException e) {
+                e.printStackTrace();
+            }
+
+            return this;
+        }
+
+
+        // set Cancelable
+        public withAnim isCancelable(boolean cancelable) {
+            setCancelable(cancelable);
+            return this;
+        }
+
+
+        // Set Title Text and Style
+        public withAnim setTitle(String string) {
+            binding.titleText.setText(string);
+            return this;
+        }
+
+        public withAnim setTitleTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.titleText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+        public withAnim setTitleTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.titleText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.titleText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setTitleTextStyle(int textStyle) {
+            if (textStyle == 1) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.titleText.setTypeface(binding.titleText.getTypeface(), Typeface.NORMAL);
+            }
+            return this;
+        }
+
+
+        // Set Description Text and Style
+        public withAnim setDescription(String descText) {
+            binding.descText.setText(descText);
+            return this;
+        }
+
+        public withAnim setDescriptionTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.descText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+        public withAnim setDescriptionTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.descText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.descText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setDescriptionTextStyle(int textStyle) {
+            // desc text style
+            if (textStyle == 1) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.descText.setTypeface(binding.descText.getTypeface(), Typeface.NORMAL);
+            }
+            return this;
+        }
+
+
+        // Set Positive Text and Style
+        public withAnim setPositiveButtonText(String string, View.OnClickListener listener) {
+            binding.positiveText.setText(string);
+            binding.positiveButton.setOnClickListener(v -> {
+                listener.onClick(v);
+                dismiss();
+            });
+            return this;
+        }
+
+        public withAnim setPositiveButtonTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.positiveText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.positiveText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setPositiveButtonTextStyle(int textStyle) {
+            if (textStyle == 1) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.positiveText.setTypeface(binding.positiveText.getTypeface(), Typeface.NORMAL);
+            }
+
+            return this;
+        }
+
+        public withAnim setPositiveButtonTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.positiveText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+
+        // Set Positive Button Style
+        public withAnim setPositiveButtonColor(int bgColorOfButton) {
+            if (bgColorOfButton != 0) {
+                try {
+                    binding.positiveButton.setCardBackgroundColor(ContextCompat.getColor(context, bgColorOfButton));
+                } catch (Resources.NotFoundException e) {
+                    binding.positiveButton.setCardBackgroundColor(bgColorOfButton);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setPositiveButtonRadius(int radiusOfButton) {
+            if (radiusOfButton != 0) {
+                binding.positiveButton.setRadius(radiusOfButton * 4);
+            }
+            return this;
+        }
+
+        public withAnim setPositiveButtonBorderColor(int borderColor) {
+            if (borderColor != 0) {
+                try {
+                    binding.positiveButton.setStrokeColor(ContextCompat.getColor(context, borderColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.positiveButton.setStrokeColor(borderColor);
+                }
+            } else {
+                binding.positiveButton.setStrokeColor(Color.TRANSPARENT);
+            }
+            return this;
+        }
+
+        public withAnim setPositiveButtonBorderWidth(int borderWidth) {
+            if (borderWidth != 0) {
+                binding.positiveButton.setStrokeWidth(borderWidth);
+            }
+            return this;
+        }
+
+
+        // Set Negative Text and Style
+        public withAnim setNegativeButtonText(String string, View.OnClickListener listener) {
+            binding.negativeText.setText(string);
+            binding.negativeButton.setOnClickListener(v -> {
+                listener.onClick(v);
+                dismiss();
+            });
+            return this;
+        }
+
+        public withAnim setNegativeButtonTextColor(int textColor) {
+            if (textColor != 0) {
+                try {
+                    binding.negativeText.setTextColor(ContextCompat.getColor(context, textColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.negativeText.setTextColor(textColor);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setNegativeButtonTextStyle(int textStyle) {
+            if (textStyle == 1) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.NORMAL);
+            } else if (textStyle == 2) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.BOLD);
+            } else if (textStyle == 3) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.ITALIC);
+            } else if (textStyle == 4) {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.BOLD_ITALIC);
+            } else {
+                binding.negativeText.setTypeface(binding.negativeText.getTypeface(), Typeface.NORMAL);
+            }
+            return this;
+        }
+
+        public withAnim setNegativeButtonTextSize(int textSizeInSP) {
+            if (textSizeInSP != 0) {
+                binding.negativeText.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSizeInSP);
+            }
+            return this;
+        }
+
+
+        // Set Negative Button Style
+        public withAnim setNegativeButtonColor(int bgColorOfButton) {
+            if (bgColorOfButton != 0) {
+                try {
+                    binding.negativeButton.setCardBackgroundColor(ContextCompat.getColor(context, bgColorOfButton));
+                } catch (Resources.NotFoundException e) {
+                    binding.negativeButton.setCardBackgroundColor(bgColorOfButton);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setNegativeButtonRadius(int radiusOfButton) {
+            if (radiusOfButton != 0) {
+                binding.negativeButton.setRadius(radiusOfButton * 4);
+
+            }
+            return this;
+        }
+
+        public withAnim setNegativeButtonBorderColor(int borderColor) {
+            if (borderColor != 0) {
+                try {
+                    binding.negativeButton.setStrokeColor(ContextCompat.getColor(context, borderColor));
+                } catch (Resources.NotFoundException e) {
+                    binding.negativeButton.setStrokeColor(borderColor);
+                }
+            } else {
+                binding.negativeButton.setStrokeColor(Color.TRANSPARENT);
+            }
+            return this;
+        }
+
+        public withAnim setNegativeButtonBorderWidth(int borderWidth) {
+            if (borderWidth != 0) {
+                binding.negativeButton.setStrokeWidth(borderWidth);
+            }
+            return this;
+        }
+
+
+        // hide options
+        public withAnim hideCloseIcon(boolean bool) {
+            if (bool) {
+                binding.closeIcon.setVisibility(View.GONE);
+            } else {
+                binding.closeIcon.setVisibility(View.VISIBLE);
+            }
+            return this;
+
+        }
+
+        public withAnim hideTitle(boolean bool) {
+            if (bool) {
+                binding.titleText.setVisibility(View.GONE);
+            } else {
+                binding.titleText.setVisibility(View.VISIBLE);
+            }
+            return this;
+        }
+
+        public withAnim hideDescription(boolean bool) {
+            if (bool) {
+                binding.descText.setVisibility(View.GONE);
+            } else {
+                binding.descText.setVisibility(View.VISIBLE);
+            }
+            return this;
+        }
+
+        public withAnim hidePositiveButton(boolean bool) {
+            if (bool) {
+                binding.positiveButton.setVisibility(View.GONE);
+                binding.padding4.setVisibility(View.GONE);
+            } else {
+                binding.positiveButton.setVisibility(View.VISIBLE);
+            }
+            return this;
+        }
+
+        public withAnim hideNegativeButton(boolean bool) {
+
+            if (bool) {
+                binding.negativeButton.setVisibility(View.GONE);
+                binding.padding4.setVisibility(View.GONE);
+            } else {
+                binding.negativeButton.setVisibility(View.VISIBLE);
+            }
+
+
+            return this;
+
+        }
+
+
+        // close icon
+        public withAnim setCloseIconListener(View.OnClickListener listener) {
+            binding.closeIcon.setOnClickListener(v -> {
+                listener.onClick(v);
+                dismiss();
+            });
+            return this;
+        }
+
+        public withAnim setCloseIcon(int iconResID) {
+            if (iconResID != 0) {
+                try {
+                    binding.closeIcon.setImageResource(iconResID);
+                } catch (Resources.NotFoundException e) {
+                    binding.closeIcon.setImageResource(R.drawable.cute_dialog_close_icon);
+                }
+            }
+            return this;
+        }
+
+        public withAnim setCloseIconSize(int sizeInDP) {
+            if (sizeInDP != 0) {
+                binding.closeIcon.getLayoutParams().height = sizeInDP * 2;
+                binding.closeIcon.getLayoutParams().width = sizeInDP * 2;
+                binding.closeIcon.requestLayout();
+            }
+            return this;
+        }
+
+        public withAnim setCloseIconColor(int colorOfIcon) {
             if (colorOfIcon != 0) {
                 try {
                     binding.closeIcon.setColorFilter(ContextCompat.getColor(context, colorOfIcon));
