--- CuteDialog/src/main/java/io/github/cutelibs/cutedialog/Enums.java
@@ -1,15 +1,0 @@
-package io.github.cutelibs.cutedialog;
-
-public class Enums {
-    // Enums
-    public static final int POSITION_CENTER = 1;
-    public static final int POSITION_TOP = 2;
-    public static final int POSITION_BOTTOM = 3;
-    public static final int STYLE_NORMAL = 1;
-    public static final int STYLE_BOLD = 2;
-    public static final int STYLE_ITALIC = 3;
-    public static final int STYLE_BOLD_ITALIC = 4;
-    public static final int HEADER_ICON = 1;
-    public static final int HEADER_IMAGE = 2;
-    public static final int HEADER_ANIMATION = 3;
-}
